Use Chef::Provider::Service::Init::Debian as provider
Users report MySQL 5.6.x not starting up on the images we created.
While we have been unable to reproduce this, we think that it is
due to the service provide not registering MySQL as such.

Previously we used Ubuntu packages, but now are are using packages from
mysql.com, which appears to support both Debian and Ubuntu.